+++ CHANGE_LOG.md
@@ -6,11 +6,14 @@
 * Added Drop and Pick fishes with variants config
 * Added block and item tags for all fish plaques
 * Added more spawn control by tags
+* Added rewards for fish collectors advancements
 * All Fish now drop item with their variants (Configurated)
 * All Fish picked up in the bucket will display their variants (Configurated)
 * All Fish loot tables now generate with their variants (Configurated)
 * Update entity movement to matches block facing when break fish plaque
 * Fixed 1 pixel error on Wrecker Bucket
 * Fixed Fish Plaque block data update from command
+* Fixed feeding fish with worms not working properly on server side
+* Fixed master fish collectors advancement not working
 * Tadpole Update :D
 * Remove unused fish plaque item texture